# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Civic Co-Pilot API - Civic Grievance and Scheme Recommendation System
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface MessageResponse {
  message: string;
}

export type RegisterRequestRole =
  (typeof RegisterRequestRole)[keyof typeof RegisterRequestRole];

export const RegisterRequestRole = {
  STUDENT: "STUDENT",
  CITIZEN: "CITIZEN",
  ADMIN: "ADMIN",
} as const;

export interface RegisterRequest {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role: RegisterRequestRole;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  STUDENT: "STUDENT",
  CITIZEN: "CITIZEN",
  ADMIN: "ADMIN",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export interface SubmitComplaintRequest {
  title: string;
  description: string;
}

export type ComplaintCategory =
  (typeof ComplaintCategory)[keyof typeof ComplaintCategory];

export const ComplaintCategory = {
  FINANCE: "FINANCE",
  INFRASTRUCTURE: "INFRASTRUCTURE",
  EDUCATION: "EDUCATION",
  HEALTH: "HEALTH",
  GENERAL: "GENERAL",
  HOUSING: "HOUSING",
  EMPLOYMENT: "EMPLOYMENT",
  ENVIRONMENT: "ENVIRONMENT",
} as const;

export type ComplaintPriority =
  (typeof ComplaintPriority)[keyof typeof ComplaintPriority];

export const ComplaintPriority = {
  LOW: "LOW",
  MEDIUM: "MEDIUM",
  HIGH: "HIGH",
  URGENT: "URGENT",
} as const;

export type ComplaintStatus =
  (typeof ComplaintStatus)[keyof typeof ComplaintStatus];

export const ComplaintStatus = {
  SUBMITTED: "SUBMITTED",
  IN_PROGRESS: "IN_PROGRESS",
  RESOLVED: "RESOLVED",
  REJECTED: "REJECTED",
} as const;

export interface Complaint {
  id: number;
  title: string;
  description: string;
  category: ComplaintCategory;
  priority: ComplaintPriority;
  department: string;
  status: ComplaintStatus;
  userId: number;
  userName?: string;
  createdAt: string;
  updatedAt: string;
}

export type StatusHistoryStatus =
  (typeof StatusHistoryStatus)[keyof typeof StatusHistoryStatus];

export const StatusHistoryStatus = {
  SUBMITTED: "SUBMITTED",
  IN_PROGRESS: "IN_PROGRESS",
  RESOLVED: "RESOLVED",
  REJECTED: "REJECTED",
} as const;

export interface StatusHistory {
  id: number;
  complaintId: number;
  status: StatusHistoryStatus;
  note?: string;
  changedAt: string;
}

export interface Scheme {
  id: number;
  name: string;
  description: string;
  category: string;
  eligibleRoles: string[];
  benefits: string;
  howToApply: string;
}

export type ComplaintDetail = Complaint & {
  statusHistory: StatusHistory[];
  recommendedSchemes: Scheme[];
};

export type UpdateStatusRequestStatus =
  (typeof UpdateStatusRequestStatus)[keyof typeof UpdateStatusRequestStatus];

export const UpdateStatusRequestStatus = {
  SUBMITTED: "SUBMITTED",
  IN_PROGRESS: "IN_PROGRESS",
  RESOLVED: "RESOLVED",
  REJECTED: "REJECTED",
} as const;

export interface UpdateStatusRequest {
  status: UpdateStatusRequestStatus;
  note?: string;
}

export type RecommendRequestUserRole =
  (typeof RecommendRequestUserRole)[keyof typeof RecommendRequestUserRole];

export const RecommendRequestUserRole = {
  STUDENT: "STUDENT",
  CITIZEN: "CITIZEN",
  ADMIN: "ADMIN",
} as const;

export interface RecommendRequest {
  complaintText: string;
  userRole: RecommendRequestUserRole;
}

export interface RecommendResponse {
  category: string;
  schemes: Scheme[];
}

export type AdminStatsByStatus = {
  SUBMITTED: number;
  IN_PROGRESS: number;
  RESOLVED: number;
  REJECTED: number;
};

export type AdminStatsByCategory = { [key: string]: number };

export type AdminStatsByPriority = {
  LOW: number;
  MEDIUM: number;
  HIGH: number;
  URGENT: number;
};

export interface AdminStats {
  totalComplaints: number;
  byStatus: AdminStatsByStatus;
  byCategory: AdminStatsByCategory;
  byPriority: AdminStatsByPriority;
  totalUsers: number;
  resolvedToday: number;
}

export type ListComplaintsParams = {
  status?: ListComplaintsStatus;
  category?: string;
};

export type ListComplaintsStatus =
  (typeof ListComplaintsStatus)[keyof typeof ListComplaintsStatus];

export const ListComplaintsStatus = {
  SUBMITTED: "SUBMITTED",
  IN_PROGRESS: "IN_PROGRESS",
  RESOLVED: "RESOLVED",
  REJECTED: "REJECTED",
} as const;

export type ListSchemesParams = {
  category?: string;
  role?: ListSchemesRole;
};

export type ListSchemesRole =
  (typeof ListSchemesRole)[keyof typeof ListSchemesRole];

export const ListSchemesRole = {
  STUDENT: "STUDENT",
  CITIZEN: "CITIZEN",
} as const;

export type AdminListComplaintsParams = {
  status?: AdminListComplaintsStatus;
  priority?: AdminListComplaintsPriority;
  department?: string;
};

export type AdminListComplaintsStatus =
  (typeof AdminListComplaintsStatus)[keyof typeof AdminListComplaintsStatus];

export const AdminListComplaintsStatus = {
  SUBMITTED: "SUBMITTED",
  IN_PROGRESS: "IN_PROGRESS",
  RESOLVED: "RESOLVED",
  REJECTED: "REJECTED",
} as const;

export type AdminListComplaintsPriority =
  (typeof AdminListComplaintsPriority)[keyof typeof AdminListComplaintsPriority];

export const AdminListComplaintsPriority = {
  LOW: "LOW",
  MEDIUM: "MEDIUM",
  HIGH: "HIGH",
  URGENT: "URGENT",
} as const;